requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain_text_splitters
lxml
langchain_openai
chromadb
langchain-chroma
faiss-cpu
sentence_transformers
langchain_huggingface
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
streamlit